Lots of little fixes to get simulation running with MICE imputed data locally

+++ minos/modules/ageing.py
@@ -57,13 +57,13 @@ class Ageing(Base):
         # update children age chains.
         population = self.update_child_ages(population)
 
-        population['nkids'] = population['nkids'].astype(float)
+        population['nkids'] = population['nkids'].astype(int)
 
         # update new population.
         logging.info(f"Aged population to year {event.time.year}")
 
         #if type(population['nkids'][0]) == float:
-        population['nkids'] = population['nkids'].astype(float)
+        population['nkids'] = population['nkids'].astype(int)
         self.population_view.update(population[['age', 'time', 'child_ages', 'nkids']])
